style(FR-3513): render BAIModal's loading body with the shared BAISkeleton
BAIModal's loading prop rendered a bare Astryx <Skeleton height={120}/> — a
single flat box — while every other loading surface in the app uses the
antd-parity multi-part shimmer (title + paragraph rows). Move that shimmer
from react/src/components/astryx-bui/BAISkeletonAstryx.tsx into
backend.ai-ui as BAISkeleton (BUI cannot import from react/src), keep the
old module as a type/default re-export so the 74 existing call sites stay
untouched, and render <BAISkeleton /> as the modal loading body.
