+++ packages/backend.ai-ui/src/components/index.ts
@@ -78,6 +78,12 @@ export type {
 } from './BAINotificationItem';
 export { default as BAIModal } from './BAIModal';
 export type { BAIModalProps } from './BAIModal';
+export { default as BAISkeleton } from './BAISkeleton';
+export type {
+  BAISkeletonProps,
+  BAISkeletonSize,
+  BAISkeletonVariant,
+} from './BAISkeleton';
 export { default as BAIBulkErrorModal } from './BAIBulkErrorModal';
 export type { BAIBulkErrorModalProps } from './BAIBulkErrorModal';
 export { default as BAIDeleteConfirmModal } from './BAIDeleteConfirmModal';
